requirements.txt
rasa-sdk
psycopg2-binary
requests
sqlalchemy<2.0
PyYAML
pytest
langdetect
